Fix 'setReward' error on boss raids, issue #36
When spawning a boss monster during a raid, the server would output an
error on the console. This was due to the fact that some raid xml files
were calling another lua script. Since this lua script was not loaded
alongside the 'global' scripts during startup, it did not contain the
function setReward() defined on [rewardboss.lua]. Because the method
setReward() was already being called in [monsters.lua], it was chosen to
delete all of the lua scripts and fix the xml files with the appropriate
tags.