fix: possible players insert conflict id / name
As we have moved the account and players creation to the migrations folder,
if the user install first the web site, there is a chance that the site
will create some account/player that can conflict with ours. To avoid
that the 'GOD' account is now created together with the database schema
and the sample players was removed from the code due to some sites like
Myaac will create them. But if you need to create those players, the
code was kept inside the migrations file.

The user '1' was removed from the code to give place to the 'GOD'
account, so the servers will now start with a single user / admin
account.

Signed-off-by: Renato Foot <rcostallat@daitangroup.com>