Update creature.lua
It is not my proposal to remove it, the system is good if it works, I don't know much about programming, so I didn't fix it. But that corrected the error.

With that code, when you have White Skull and let the monsters attack you, you give it to exit and re-enter as in 20 seconds you will be kicked with white skull.