New door system, 100% tested and working.
Now everything will be organized in just one table, containing each type of door (key locked / key unlocked / custom door / level door / quest door).
Easy to understand, only with openDoor and closedDoor to configure, just register and use, it is not necessary to add the door ID in action.xml and movements.xml.

It fixes the task 3 of issue #644
Regards!