Tested and fixed the bugs in the First Dragon quest.
It just remains bugs to bring the final dragon to First Dragon, soon it will be fixed, the script is complex.

Bugs fixeds:

It was not possible to enter the boss TPS after finishing tasks.
Task Life for boss Zorvorax was not functional, the movements were missing to collect the accesses.

Task Power for boss Tazhadur was not functional, corrected the access teleport.

Treasure for boss Kalyassa was not functional, it was not possible to enter the TP after collecting the 5 rewards and the achievement was given after collecting the first, and not after collecting the 20 (corrected). The achievement did not exist either, it was placed.

Fixed access to the Dragon Lair with the Scale of Gelidrazah, Horn of Kalyassa, Bones of Zorvorax and Tooth of Tazhadur items when using them in front of the gate.