Save attributes and characters in a certain period of time
The script originally existed, but for some reason I don't know, it was removed.
What currently happens is that if the server crashes or is closed without saving it will return to the last server save, as there is no script that saves every time, this can be configured for every 30 minutes or one hour, according to the particularity of each user, but I recommend that it is every 30 minutes.