Continued work in bestiary and charms (#964)
* Bestiary update and charms fix

* Continued work in bestiary and charms

Now it blacks out the monster when you dont have it unlocked yet. Charms can be unlocked. There is a charm expansion method aswell, which someone should link to the gamestore (player:setCharmRuneSlotExpansion(true/false)). There is also a method that you can use to get the active charm for the monster(if any): player:getCharmFromTarget(target)

TODO add the methods to trigger the charm runes
TODO add information related to monster locations in bestiary
TODO add information of loot form special event (this should be fixed in creature loots and then parsed to the bestiary)

Special thanks to Rick and lBaah :)

* Make use of assets indexes on the power of two operations

* Change table name

* Update credits

* Update bestiary.lua

* Update bestiary.lua

* [Bestiary] Removed DB dependancy and switched to storages. Added new commands aswell

Did stuff in the title and also implemented the cyclopedia glow button when bestiary entry has new update.

* Fix crash on monsters with occurencies different than expected

Co-authored-by: Eduardo Dantas <eduardo.dantas@hotmail.com.br>