Oressa fixes and tiles vocation back
Corrected Oressa NPC (now as global)
Now, after choosing the vocation with Oressa, it will no longer be possible to climb the ladder,
After passing the door it is not possible to step on the tiles that give access to the door.
After leaving the chest room, it is no longer possible to return.
Everything developed as it is in the global.

Added texts on some downport signs.