Dawnport fixes and development
Added dawnport books and signs
Added dawnport quest full
Revised dawnport npcs (as well as all revised lines)
Npc Plunderpurse now limits the deposit to 1000 gps
Npc Wentworth limits the deposit to 1000 gps for lvl 8 and 2000 gps for lvl 9 to 20
Revised npc Black Bert
Oressa will not remove the legion helmet
Upon passing the vocation door, you will remove the items (with the exception of the legion helmet and tools)

Added function to load the Book and Sign table
Refactored the function of loading the action table, it is now possible to add an action without having to check the itemId, setting the variable "itemId = false", thus allowing to add the same action in several different ids and in different positions. Notice that the action is added to the top item.