Vocations fixes
Created 4 vocations for dawnport.
The dawnport vocations have rooker status, that is, they have a fixed gain of health/mana/cap and other attributes up to level 8. Therefore, proper vocations for dawnport are necessary.

As a result of this change, an enum was created with vocation and clientids ids in the tables folder in libs.
Several calls to vocations were changed from "ids" to "clientid", since before it was bugged and only allowed use with the vocation where the id was set. That way, it will also be possible to use the calls like "elite knight", "royal paladin" and etc ...

Several minor fixes were made, code refactoring and improvement.

Table of additional maps moved to the folder "tables" and the script attached to startup.lua.
Towns table moved to the "tables" folder.