Small fixes in dawnport (thanks for @heitortadeu per the reports)
Correction in oressa (problem when choosing vocation)
Fixed an event that teleports the player to the temple five minutes after reaching level 20)
This was a problem because, if the player left dawnport before five minutes, when the event ran, he would return to dawnport and be stuck.
Thus, I added the stopEvent on Oressa, after choosing the vocation.

Correction in migrations.