Fixed Grizzly Adams not buying anything from the player.
Grizzly wasn't buying anything the player tried to sell and was throwing an error.
I was able to fix this issue but this script probably needs a remake, It's messy and probably outdated, I am currently not skilled enough to do it but this at least makes him able to trade again.