Fix Load Player by Name
- Fix load player by name that was not getting bonus_reroll field, so
in place of getting each field, we now get all fields to avoid loading
by name or by ID to have differents fields on the query.

This fixes an error when sending parcel that use the load player by name
and was complaining that 'bonus_reroll' was missing from the
result of the query.

Signed-off-by: Renato Foot <costallat@hotmail.com>