Fix: Daily Reward (#1773)
Resolves #1178

Now is properly showing the timer when the next reward is available to be taken again or the time left to take the daily reward on the current day.

Instant daily token feat.

Show the right potions/runes that the player can use. Ex: knight only can take 'mana potion' and all health potion.

Regeneration bonus is working properly.

Resting area icon is showing the right bonus activated.

On login, if the player did not get his daily reward on the current day, the 'reward' icon will have the golden mark around it. image

ToDo:

Migrate to cpp. (If someone is available to do it I would be grateful)
Anything that is not working as it should be PLEASE leave a comment.

NOTE: this system uses the server save time, so if you are testing it and you don't use the server save system, you might have some problems on testing it.