Fixed urmahlullu to current Tibia version (#2003)
- [X] Fix changes Urmahlullu
- [X] Fix lever to boss (as daily boss)

```
The fight has five stages. First, you must defeat Urmahlullu the Immaculate . There are two spots viable for trapping it.
One of them is in the south-eastern part of the cave, one is in the north-eastern part, near the teleport.
Once you defeat this phase, Wildness of Urmahlullu will spawn. Keep in mind that you have around a minute to kill it,
otherwise it will revert to previous form. Once killed, it will become Urmahlullu the Tamed . Next one is
Wisdom of Urmahlullu . Just like Wildness form, it will rever to its previous form if not killed on time.
This form is immune to death damage. The last form is calledUrmahlullu The Weakened and it will drop 
the Ring of Secret Thoughts .
```
Source: https://www.tibiaroyal.com/en/guide/quest/280/kilmaresh-quest/spoiler