Fix loadLuaMapBook function (#2309)
Fixes the loadLuaMapBook function not working properly. Is generating containers on the map when book item need to be generated at a map position and when the book item need to be in a container it isnt generated inside. Only works on existing already on map items.

Fixed the issues, now generate new book items at map and inside containers, and still working on existing map items.
Moved the loading console message inside the function.
Merged the BookTable and DocumentTable. IMHO they are like the same, items that can be on map or inside a container that when are used they show a window with a readable text.
Renamed function name and some changes on table variables.